+++ views/index.html
@@ -8,8 +8,11 @@
     <meta name = "Keywords" content = "Dice, Roll, D&D">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <script type="module" src="/scripts/dat.gui.module.js"></script>
+    <script type="module" src="/public/roll3DDice.js"></script>
+    <script type="module" src="/public/draw2dDiceAndError.js"></script>
 </head>
 <body>
+<audio src="/resources/Sen no Kiseki II OST - Chasing After the Mountain Hare.mp3" controls autoplay loop>Your browser does not support audio</audio>
 <div class = "heading">
     <h1>
         <div id = "title">
